Introduce simulator-kite api and kernel skeleton
The classes added in this commit constitute an overview of the
conceptual model of this project.

The Kernel is able to load a program from a file but not run it.

The Kernel class uses a Simulator api, which hides the cpu
components and provides methods to read/write to these components.
Kernel is able to load a program from a file (using code
provided in pyrisc project). It also has a start method, which
represents the kernel main loop but is not fully implemented.
Main kernel concepts are introduced: process, cpu_context, scheduler.