Separate parsing from load_process_from_file
Introduce method parse_cpu_context_from_file and move it away from
kernel class.
I will see how load_process_from_file will develop, but I'm not sure if it
will be useful in the future.
Parsing ELF files will occur only in two cases: when kernel starts and in
execve. But we don't want to return a new process on execve. So maybe
parsing is the only abstraction we want.